Animal/Outreach/Vet apps
outreach and animal applications submit to both tables correctly.  I
added password check (make sure the passwords are the same) to the
animal file.  It works correctly but refreshes all fields after the
error (can't have that).  I put some code to get started on the Vet app
but not sure if the fields on the app are correct so did not begin to
commit those.